+++ fastchart_arginfo.h
@@ -1,5 +1,5 @@
 /* This is a generated file, edit the .stub.php file instead.
- * Stub hash: 52d62b3b4c4407a4646884d7de448ce3270c0a2b */
+ * Stub hash: cbdf2f65ee9f5ef5f044bba8b3c044a0fa8e49ef */
 
 ZEND_BEGIN_ARG_INFO_EX(arginfo_class_FastChart_Chart___construct, 0, 0, 0)
 	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, width, IS_LONG, 1, "null")