+++ php_fastchart.h
@@ -21,7 +21,7 @@
 
 #include "fastchart_graph.h"
 
-#define PHP_FASTCHART_VERSION "1.7.1"
+#define PHP_FASTCHART_VERSION "1.7.2"
 
 extern zend_module_entry fastchart_module_entry;
 #define phpext_fastchart_ptr &fastchart_module_entry